# core-common-security-vault-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-security-vault
  description: Common Platform Security Vault - Secure Credential Storage
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/vault
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
  /api/v1/environment-types/{id}:
    get:
      tags:
        - Environment Types
      summary: Get an environment type by ID
      operationId: getEnvironmentTypeById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnvironmentTypeDTO'
    put:
      tags:
        - Environment Types
      summary: Update an existing environment type
      operationId: updateEnvironmentType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentTypeDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnvironmentTypeDTO'
    delete:
      tags:
        - Environment Types
      summary: Delete an environment type
      operationId: deleteEnvironmentType
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: No Content
...
  /api/v1/environment-types:
    post:
      tags:
        - Environment Types
      summary: Create a new environment type
      operationId: createEnvironmentType
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnvironmentTypeDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnvironmentTypeDTO'
  /api/v1/environment-types/filter:
    post:
      tags:
        - Environment Types
      summary: Filter environment types
      operationId: filterEnvironmentTypes
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestEnvironmentTypeDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseEnvironmentTypeDTO'
...
components:
  schemas:
    EnvironmentTypeDTO:
      required:
        - code
        - name
      type: object
      properties:
        id:
          type: string
          description: Environment type ID
          format: uuid
          readOnly: true
        code:
          maxLength: 50
          minLength: 2
          type: string
          description: Unique code for the environment type
          example: PRODUCTION
        name:
          maxLength: 100
          minLength: 2
          type: string
          description: Name of the environment type
          example: Production
        description:
          type: string
          description: Description of the environment type
          example: Production environment
        active:
          type: boolean
          description: Whether the environment type is active
          default: true
        version:
          type: integer
          description: Version for optimistic locking
          format: int64
          readOnly: true
        createdAt:
          type: string
          description: Creation timestamp
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          description: Last update timestamp
          format: date-time
          readOnly: true
      description: Environment Type information
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
    FilterRequestEnvironmentTypeDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/EnvironmentTypeDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponseEnvironmentTypeDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/EnvironmentTypeDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...